Rename file name for PurchaserInfo+HybridAdditions, update DEVELOPMENT.md for single pod
