+++ DEVELOPMENT.md
@@ -5,14 +5,13 @@
 In `ios/PurchasesHybridCommon/Podfile` replace:
 
 ```
-  pod 'Purchases', '3.5.1'
+  pod 'RevenueCat', '4.0.0'
 ```
 
 with:
 
 ```
-  pod 'Purchases', :path => '~/Development/repos/ios/purchases-ios'
-  pod 'PurchasesCoreSwift', :path => '~/Development/repos/ios/purchases-ios'
+  pod 'RevenueCat', :path => '~/Development/repos/ios/purchases-ios'
 ```
 
 Remember to change `'~/Development/repos/ios/purchases-ios'` with your local purchases-ios path.
